split sdk package into dsnp and content and document
